portfolio-rag-assistant/requirements.txt
streamlit==1.32.0
chromadb==1.5.7
groq
python-dotenv==1.0.1
